Extracting TestInstanceInfoContext to a separate class to be used across the tests
